Add loan form features
