+++ app/assets/stylesheets/application.scss
@@ -19,6 +19,7 @@
  @import 'devise';
  @import 'painel';
 
+ @import 'loan';
  
 :root{
   --shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
