+++ app/assets/stylesheets/loan.scss
@@ -1,3 +1,9 @@
 // Place all the styles related to the loan controller here.
 // They will automatically be included in application.css.
 // You can use Sass (SCSS) here: https://sass-lang.com/
+
+#loan{
+  .loan-form{
+    padding: 20px 10px 10px 10px;
+  }
+}
