+++ app/views/loan/_form.html.erb
@@ -1,17 +1,29 @@
-<%= form_with url: "/", method: :get do |f| %>
-  <%= f.hidden_field :monthly_rate, value:@loan.monthly_rate %>
-  <label for="">Valor do Emprestimo</label>
-  <div class="input-group mb-3">
-    <div class="input-group-prepend">
-      <span class="input-group-text">R$</span>
+<div class="loan-form">
+  <%= form_with url: "/", method: :get do |f| %>
+    <label for="">Valor do Emprestimo</label>
+    <div class="input-group mb-3">
+      <div class="input-group-prepend">
+        <span class="input-group-text">R$</span>
+      </div>
+      <%= f.number_field :value, class:"form-control" %>
+      <div class="input-group-append">
+        <span class="input-group-text">.00</span>
+      </div>
     </div>
-    <%= f.text_field :value, class:"form-control" %>
-    <div class="input-group-append">
-      <span class="input-group-text">.00</span>
+    <div class="form-group">
+      <label for="">Taxa <small> %a.m.</small></label>
+      <%= f.text_field :monthly_rate, class:"form-control", disabled: true, value:(@loan.monthly_rate*100 )%>
     </div>
-  </div>
-  <div class="form-group">
-    <%= f.select :installments,(1..60).map{|n| n},{}, {class:"custom-select"}%>
-  </div>
-  <button type="submit" class="btn btn-primary">Submit</button>
-<%end%>
+    <div class="form-group">
+      <label for="">Taxa</label>
+      <%= f.select :installments,(1..60).map{|n| n},{}, {class:"custom-select"}%>
+    </div>
+    <div class="form-check pb-2">
+      <input class="form-check-input" type="checkbox"  id="confirm">
+      <label class="form-check-label" for="confirm">
+        Aceito os termos da solicitação de credito
+      </label>
+    </div>
+    <input disabled type="submit" class="btn btn-primary" value="Proxima etapa"/>
+  <%end%>
+</div>
