+++ app/views/loan/new.html.erb
@@ -1,7 +1,6 @@
-<div id="loan" class="cont">
-  <div class="loan-form">
+<div id="loan" class="cont p-3">
+  <h3>Nova solicitação de crédito</h3>
+  <h6 class="pl-3 border-bottom"><i><b>Etapa</b> 1/2 </i></h6>
+  <%= render 'preview' %>
   <%= render 'form' %>
-  </div>
-  <div class="loan-preview">
-  </div>
 </div>
